--- remote_attestation/rust/src/attestation/attested_session.rs
@@ -1,45 +1,0 @@
-//
-// Copyright 2021 The Project Oak Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-
-use crate::{crypto::AeadEncryptor, proto::EncryptedData};
-use anyhow::Context;
-
-pub struct AttestedSession {
-    /// Utility object for decrypting and encrypting messages using a Diffie-Hellman session key.
-    encryptor: AeadEncryptor,
-}
-
-impl AttestedSession {
-    pub fn new(encryptor: AeadEncryptor) -> Self {
-        Self { encryptor }
-    }
-
-    pub fn encrypt(&mut self, data: &[u8]) -> anyhow::Result<EncryptedData> {
-        let encrypted_data = self
-            .encryptor
-            .encrypt(data)
-            .context("Couldn't encrypt data")?;
-        Ok(encrypted_data)
-    }
-
-    pub fn decrypt(&mut self, data: &EncryptedData) -> anyhow::Result<Vec<u8>> {
-        let decrypted_data = self
-            .encryptor
-            .decrypt(data)
-            .context("Couldn't decrypt message")?;
-        Ok(decrypted_data)
-    }
-}
